requirements.txt
Django==4.2.7
python-docx==1.1.0
PyPDF2==3.0.1
chromadb==0.4.18
openai==1.3.6
numpy>=1.21.0
sentence-transformers>=2.2.0
python-dotenv==1.0.0
django-cors-headers==4.3.1